✨ feat(Home): Animation
Animation between page
