+++ package.json
@@ -11,16 +11,20 @@
   "dependencies": {
     "@react-spring/parallax": "^9.4.5",
     "@react-spring/web": "^9.4.5",
+    "framer-motion": "^6.3.11",
     "next": "12.1.6",
+    "polished": "^4.2.2",
     "react": "18.1.0",
     "react-dom": "18.1.0",
     "react-icons": "^4.3.1",
-    "react-spring": "^9.4.5"
+    "react-spring": "^9.4.5",
+    "styled-components": "^5.3.5"
   },
   "devDependencies": {
     "@types/node": "17.0.36",
     "@types/react": "18.0.9",
     "@types/react-dom": "18.0.5",
+    "@types/styled-components": "^5.1.25",
     "autoprefixer": "^10.4.7",
     "eslint": "8.16.0",
     "eslint-config-next": "12.1.6",
