+++ src/pages/_app.tsx
@@ -1,8 +1,20 @@
-import '@/assets/css/globals.css'
-import type { AppProps } from 'next/app'
+import "@/assets/css/globals.css";
+import type { AppProps } from "next/app";
+import { AnimatePresence } from "framer-motion";
+import { useRouter } from "next/router";
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
+  const router = useRouter();
+
+  return (
+    <AnimatePresence
+      initial={false}
+      presenceAffectsLayout={false}
+      onExitComplete={() => window.scrollTo(0, 0)}
+    >
+      <Component {...pageProps} key={router.route} />
+    </AnimatePresence>
+  );
 }
 
-export default MyApp
+export default MyApp;
