+++ src/pages/projects/index.tsx
@@ -1,0 +1,15 @@
+import MainLayout from "@/layouts/main.layout";
+
+const Projects = () => {
+  return (
+    <MainLayout>
+      <section className="min-h-screen bg-slate-900">
+        <div className="container mx-auto">
+          <h1>Projects</h1>
+        </div>
+      </section>
+    </MainLayout>
+  );
+};
+
+export default Projects;
